Llama-3 works; before cleaning up cache
