+++ scripts/example_run_gcg.sh
@@ -6,10 +6,10 @@ ATTACK="gcg"
 
 python -u main.py \
     --config="./configs/${ATTACK}.py" \
-    --config.batch_size=512 \
+    --config.batch_size=256 \
     --config.num_steps=30 \
     --config.log_freq=1 \
-    --config.fixed_params=False \
+    --config.fixed_params=True \
     -- \
     --scenario "Toxicity" --behaviors 0 --system_message "llama_default" \
     --model llama-2@~/data/models/Llama-2-7b-chat-hf --verbose
