+++ src/models/openai.py
@@ -17,8 +17,9 @@ from openai.types.chat.chat_completion_token_logprob import TopLogprob
 from tenacity import retry, retry_if_not_exception_type, wait_random_exponential
 
 from src.message import Message, Role
-from src.models.base import BaseModel, Encoded, LossOutput
+from src.models.base import BaseModel, LossOutput
 from src.models.model_input import ModelInputIds, ModelInputs
+from src.models.tokenizer import GptTokenizer
 from src.servers.openai_server import Queues, standalone_server
 from src.utils.suffix import SuffixManager, build_prompt
 from src.utils.types import BatchTokenIds, TokenIds
@@ -45,82 +46,6 @@ COMPLETION_MODELS = [
 ]
 
 logger = logging.getLogger(__name__)
-
-
-class GptTokenizer:
-    def __init__(self, encoding: tiktoken.Encoding):
-        # Get the tokeniser corresponding to a specific model in the OpenAI API
-        self.encoding: tiktoken.Encoding = encoding
-        # Set interface to match HuggingFace
-        self.vocab_size = self.encoding.max_token_value + 1
-        self.bos_token_id = self.encoding.eot_token
-        self.eos_token_id = self.encoding.eot_token
-        self.pad_token_id = self.encoding.eot_token
-        self.unk_token_id = self.encoding.eot_token
-        self.eot_token = "<|endoftext|>"
-
-    def __len__(self):
-        return self.vocab_size
-
-    def __call__(
-        self,
-        text: str,
-        return_tensors: Literal["list", "pt", "np"] = "list",
-        **kwargs,
-    ):
-        _ = kwargs  # unused
-        if text is None:
-            return Encoded([])
-
-        if isinstance(text, list):
-            # Encode all special tokens as normal text
-            _ids = self.encoding.encode_batch(text, disallowed_special=())
-            max_len = max(len(i) for i in _ids)
-            input_ids = np.zeros((len(_ids), max_len), dtype=np.int64)
-            input_ids += self.pad_token_id
-            for i, _id in enumerate(_ids):
-                input_ids[i, : len(_id)] = _id
-        else:
-            input_ids = self.encoding.encode(text, disallowed_special=())
-            input_ids = np.array(input_ids, dtype=np.int64)
-
-        if return_tensors == "pt":
-            input_ids = torch.from_numpy(input_ids)
-        elif return_tensors == "list":
-            input_ids = input_ids.tolist()
-        return Encoded(input_ids)
-
-    def _parse_ids(self, ids):
-        return ids
-
-    def decode(self, ids, **kwargs) -> str:
-        _ = kwargs  # unused
-        if isinstance(ids, torch.Tensor):
-            ids = ids.tolist()
-        if isinstance(ids, int):
-            ids = [ids]
-        assert isinstance(ids, list) and isinstance(
-            ids[0], int
-        ), f"ids must be list or int, got {type(ids)} {ids}"
-        decoded = self.encoding.decode(ids)
-        return decoded.replace(self.eot_token, "")
-
-    def batch_decode(self, ids, **kwargs) -> list[str]:
-        _ = kwargs  # unused
-        if isinstance(ids, torch.Tensor):
-            ids = ids.tolist()
-        if isinstance(ids, int):
-            ids = [[ids]]
-        if isinstance(ids, list) and isinstance(ids[0], int):
-            ids = [ids]
-        assert (
-            isinstance(ids, list)
-            and isinstance(ids[0], list)
-            and isinstance(ids[0][0], int)
-        ), f"ids must be list of list of int, got {type(ids)} {ids}"
-        decoded_list = self.encoding.decode_batch(ids)
-        decoded_list = [s.replace(self.eot_token, "") for s in decoded_list]
-        return decoded_list
 
 
 class Response:
